organized actions, fix subject add flow
