+++ web/src/reducers/core.js
@@ -1,0 +1,5 @@
+export const LOGIN_FAIL = "LOGIN_FAIL"
+export const createLoginFail = () => ({ type: LOGIN_FAIL })
+
+export const LOGIN_SUCCEED = "LOGIN_SUCCEED"
+export const createLoginSucceed = (school_id, db, token) => ({ type: LOGIN_SUCCEED, school_id, db, token })
